Dfu (#59)
* initial states for dfu

* initial states for DFU

* finished architecture for DFU class

* formatting

* removed error enum, opting for FotaError struct enum members instead ifor dfu member functions.

* removed packet_crc32 from DfuStateData, because not needed in DFU (done in crc check)

* Small fixes (packet manager added to DFUStateData), and made state machine startable by app layer.

* started writing implementation for dfu

* finished writing process_new_state function to switch states

* small change

* refactored process_new_state

* run_dfu_bootloader

* dfu bootloader rough implementation

* naming issues in fota flash fixed

* device firmware update state machine implementation.

* made changes based on review

* Set up GUI library projectgit add .!

* Create files in gui library

* Add new lines at the end of each file

* Try again ;-;

* Try no whitespace at the end??

* Remove whitespace from main and test_gui

* only format gui files

* Add new lines to rest of files

* Fix build

* Fix formatting

* Formatting

* Remove accidental files

* Fix build

---------

+++ projects/fota/inc/fota_flash.h
@@ -11,6 +11,7 @@
 
 /* Standard library Headers */
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 
 /* Inter-component Headers */
@@ -60,7 +61,7 @@
  *          FOTA_ERROR_FLASH_WRITE_OUT_OF_BOUNDS if address is out of bounds
  *          FOTA_ERROR_FLASH_WRITE_NOT_ALIGNED if data is not aligned
  */
-FotaError fota_flash_write(uintptr_t address, uint8_t *buffer, uint32_t buffer_len);
+FotaError fota_flash_write(uint32_t address, uint8_t *buffer, size_t buffer_len);
 
 /**
  * @brief   Erase some number of flash pages
@@ -79,7 +80,7 @@ FotaError fota_flash_erase(uint8_t start_page, uint8_t num_pages);
  * @return  FOTA_ERROR_SUCCESS if data is read succesfully
  *          FOTA_ERROR_FLASH_READ_FAILED if data read failed
  */
-FotaError fota_flash_read(uintptr_t address, uint8_t *buffer, uint32_t buffer_len);
+FotaError fota_flash_read(uint32_t address, uint8_t *buffer, size_t buffer_len);
 
 /**
  * @brief   Verify that the flash memory is not erased
